requirements.txt
click==8.*
requests==2.32.3
pygithub==2.3.0
numpy<2.0.0
pydantic==2.8.2
scipy==1.14.0
lxml-html-clean==0.1.1
pandas==2.2.2
tabulate==0.9.0
pyyaml==6.0.1
jupyter-cadquery @ git+https://github.com/OpenOrion/jupyter-cadquery.git
GitPython==3.1.43